modify_migrationfile

+++ db/migrate/20250109151222_create_reviews.rb
@@ -1,8 +1,8 @@
 class CreateReviews < ActiveRecord::Migration[6.1]
   def change
     create_table :reviews do |t|
-      t.integer :user_id, null: false
-      t.integer :hut_id, null: false
+      t.bigint :user_id, null: false
+      t.bigint :hut_id, null: false
       t.string :title, null: false
       t.text :body, null: false
       t.integer :rating, null: false
