+++ db/migrate/20250109151428_create_comments.rb
@@ -1,8 +1,8 @@
 class CreateComments < ActiveRecord::Migration[6.1]
   def change
     create_table :comments do |t|
-      t.integer :user_id, null: false
-      t.integer :review_id, null: false
+      t.bigint :user_id, null: false
+      t.bigint :review_id, null: false
       t.string :body, null: false
       t.timestamps
     end
